Make final_address part of hbit::Params
The hbit protocol is designed in a way that makes the use of transient
identities mandatory. That is because there is currently no way for
how a Bitcoin wallet would know how to sign the redeem/refund tx.

As such, the transient secret key and the final identity are part
of the _local_ parameters of the protocol.

For some obscure reason, we cannot use p2pkh addresses in the arbitrary
implementation because the serialization doesn't roundtrip. Hence we
switch to p2wpkh.