Add commands to nectar for creating and signing redeem/refund tx
These commands take the swap-id, the HTLC location and in case of redeem
also the secret and produce a signed transaction for the particular ledger
that just needs to be broadcasted in order to trigger this action.

This can be useful in cases where nectar's swap execution gets stuck
and one just wants to manually trigger a specific action.