Add sub-commands to cnd that aid manual recovery
For the hbit protocol, we create and sign the transaction for redeeming
and refunding within the node. We can offer sub-commands that produce
exactly these transactions for a swap.

For the herc20 protocol, we don't create and sign these transactions.
However, manually redeeming / refunding those is fairly trivial
through MyEtherWallet as long as you have the secret. We offer a sub-
command for printing the secret in case we are Alice.