Fix the test
Alice was set to "alpha funded", time was moved forward to assume that
Bob acted but this was not reflected in Alice state.

Alice next state is "beta funded" so the test should simulate that,
indeed, Bob waits for the last minute but also his (beta) transaction is
mined and reached finality so we can reflect that on Alice state, by
changing it to "beta funded".