Fix mess of hbit::Params and swap::hbit::Params
We can collapse those two modules together and thereby simplify
the usage and imports in several places.

Unfortunately, we cannot (yet) completely get rid of SharedParams
(and simply merge the structs) because we also use the SharedParams
struct in places where we don't have access to the private information
like the secret key.