Replace loop-based retry with backoff crate
For now, the functionality is pretty much the same because we use
a constant backoff of 1 second. Previously we used to spin as fast
as possible which would yield _a lot_ of errors in case there were
any.
Using the backoff crate solves this and also allows us to easily
change the backoff strategy.