Remove `Option<T>` where possible
Previously, some of the fields in the database schema where nullable.
However, with the introduction of orders and the removal of the endpoints
for directly creating swaps, we always have all the information
available to insert the complete swap. As such, we can simplify some
of the DB code by removing these `Option`s.