Only trigger mergify if none of the build checks fail
Recently [1], we've expanded our CI build checks from 3 to 4. Maintaining
this list of checks became cumbersome and as such, we tried to
optimize this [2].

Unfortunately, the solution in [2] didn't work because it triggers
as soon as _any_ of the "build" checks succeed.

This patch uses double-negation to make sure we only trigger mergify
if none of the "build" checks fail. Consequently, all of them need
to succeed for mergify to consider this rule matching.

[1]: https://github.com/comit-network/comit-rs/pull/3364
[2]: https://github.com/comit-network/comit-rs/pull/3364